+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/EventHookCallback.java
@@ -18,15 +18,17 @@
 
 package org.apache.skywalking.oap.server.core.alarm.provider;
 
+import java.io.IOException;
 import org.apache.skywalking.apm.network.event.v3.Event;
 import org.apache.skywalking.apm.network.event.v3.Source;
 import org.apache.skywalking.apm.network.event.v3.Type;
 import org.apache.skywalking.oap.server.analyzer.event.EventAnalyzerModule;
 import org.apache.skywalking.oap.server.analyzer.event.EventAnalyzerService;
+import org.apache.skywalking.oap.server.core.CoreModule;
 import org.apache.skywalking.oap.server.core.alarm.AlarmCallback;
 import org.apache.skywalking.oap.server.core.alarm.AlarmMessage;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
-import org.apache.skywalking.oap.server.core.analysis.Layer;
+import org.apache.skywalking.oap.server.core.query.MetadataQueryService;
 import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 
@@ -41,22 +43,36 @@ import java.util.UUID;
 public class EventHookCallback implements AlarmCallback {
 
     private final ModuleManager manager;
+    private MetadataQueryService metadataQueryService;
+
+    private MetadataQueryService getMetadataQueryService() {
+        if (metadataQueryService == null) {
+            this.metadataQueryService = manager.find(CoreModule.NAME)
+                                                     .provider()
+                                                     .getService(MetadataQueryService.class);
+        }
+        return metadataQueryService;
+    }
 
     public EventHookCallback(ModuleManager manager) {
         this.manager = manager;
     }
 
     @Override
-    public void doAlarm(List<AlarmMessage> alarmMessage) {
+    public void doAlarm(List<AlarmMessage> alarmMessage) throws Exception {
         EventAnalyzerService analyzerService = manager.find(EventAnalyzerModule.NAME).provider().getService(EventAnalyzerService.class);
-        alarmMessage.forEach(a -> {
+        for (AlarmMessage a : alarmMessage) {
             for (Event event : constructCurrentEvent(a)) {
                 analyzerService.analyze(event);
             }
-        });
+        }
     }
 
-    private List<Event> constructCurrentEvent(AlarmMessage msg) {
+    private String getLayer(String serviceId) throws IOException {
+        return getMetadataQueryService().getService(serviceId).getLayers().iterator().next();
+    }
+
+    private List<Event> constructCurrentEvent(AlarmMessage msg) throws IOException {
         List<Event> events = new ArrayList<>(2);
         long now = System.currentTimeMillis();
         Event.Builder builder = Event.newBuilder()
@@ -65,8 +81,8 @@ public class EventHookCallback implements AlarmCallback {
                 .setStartTime(now - (msg.getPeriod() * 60 * 1000))
                 .setMessage(msg.getAlarmMessage())
                 .setType(Type.Error)
-                .setEndTime(now)
-                .setLayer(Layer.GENERAL.name());
+                .setEndTime(now);
+               // .setLayer(Layer.GENERAL.name());
         switch (msg.getScopeId()) {
             case DefaultScopeDefine.SERVICE :
                 IDManager.ServiceID.ServiceIDDefinition serviceIdDef = IDManager.ServiceID.analysisId(msg.getId0());
@@ -75,6 +91,7 @@ public class EventHookCallback implements AlarmCallback {
                             .setService(serviceIdDef.getName())
                             .build()
                 );
+                builder.setLayer(getLayer(msg.getId0()));
                 events.add(builder.build());
                 break;
             case DefaultScopeDefine.SERVICE_RELATION :
@@ -91,6 +108,7 @@ public class EventHookCallback implements AlarmCallback {
                                 .setService(destServiceIdDef.getName())
                                 .build()
                 ).setUuid(UUID.randomUUID().toString());
+                builder.setLayer(getLayer(sourceServiceIdDef.isReal() ? msg.getId0() : msg.getId1()));
                 events.add(builder.build());
                 break;
             case DefaultScopeDefine.SERVICE_INSTANCE :
@@ -101,6 +119,7 @@ public class EventHookCallback implements AlarmCallback {
                                 .setService(IDManager.ServiceID.analysisId(instanceIdDef.getServiceId()).getName())
                                 .build()
                 );
+                builder.setLayer(getLayer(instanceIdDef.getServiceId()));
                 events.add(builder.build());
                 break;
             case DefaultScopeDefine.SERVICE_INSTANCE_RELATION :
@@ -119,6 +138,8 @@ public class EventHookCallback implements AlarmCallback {
                                 .setService(IDManager.ServiceID.analysisId(destInstanceIdDef.getServiceId()).getName())
                                 .build()
                 ).setUuid(UUID.randomUUID().toString());
+                IDManager.ServiceID.ServiceIDDefinition srcInstanceServiceIdDef = IDManager.ServiceID.analysisId(sourceInstanceIdDef.getServiceId());
+                builder.setLayer(getLayer(srcInstanceServiceIdDef.isReal() ? sourceInstanceIdDef.getServiceId() : destInstanceIdDef.getServiceId()));
                 events.add(builder.build());
                 break;
             case DefaultScopeDefine.ENDPOINT :
@@ -129,6 +150,7 @@ public class EventHookCallback implements AlarmCallback {
                                 .setService(IDManager.ServiceID.analysisId(endpointIDDef.getServiceId()).getName())
                                 .build()
                 );
+                builder.setLayer(getLayer(endpointIDDef.getServiceId()));
                 events.add(builder.build());
                 break;
             case DefaultScopeDefine.ENDPOINT_RELATION :
@@ -147,6 +169,9 @@ public class EventHookCallback implements AlarmCallback {
                                 .setService(IDManager.ServiceID.analysisId(destEndpointIDDef.getServiceId()).getName())
                                 .build()
                 ).setUuid(UUID.randomUUID().toString());
+                IDManager.ServiceID.ServiceIDDefinition srcEndpointServiceIdDef = IDManager.ServiceID.analysisId(sourceEndpointIDDef.getServiceId());
+                IDManager.ServiceID.ServiceIDDefinition destEndpointServiceIdDef = IDManager.ServiceID.analysisId(destEndpointIDDef.getServiceId());
+                builder.setLayer(getLayer(srcEndpointServiceIdDef.isReal() ? sourceEndpointIDDef.getServiceId() : destEndpointIDDef.getServiceId()));
                 events.add(builder.build());
                 break;
         }