+++ docs/en/changes/changes.md
@@ -21,6 +21,8 @@
 * Fix table exists check in the JDBC Storage Plugin.
 * Enhance extensibility of HTTP Server library.
 * Adjust `AlarmRecord` alarmMessage column length to 512.
+* Fix `EventHookCallback` build event: build the layer from `Service's Layer`.
+* Fix `AlarmCore` doAlarm: catch exception for each callback to avoid interrupt.
 
 #### UI